+++ guillotina_elasticsearch/tests/conftest.py
@@ -1,12 +1,12 @@
 from pytest_docker_fixtures import images
 
 
-image_version_8 = "8.12.0-debian-11-r2"
-image_version_7 = "7.17.16-debian-11-r3"  # noqa
+image_version_8 = "8.12.0"
+image_version_7 = "7.17.16"  # noqa
 
 images.configure(
     name="elasticsearch",
-    full=f"bitnami/elasticsearch:{image_version_8}",
+    full=f"bitnamilegacy/elasticsearch:{image_version_8}",
     max_wait_s=90,
     env={"ELASTICSEARCH_ENABLE_SECURITY": "false", "ELASTICSEARCH_HEAP_SIZE": "1g"},
 )